move parameters to separate class
