+++ common-tools/swim-tools/src/main/java/org/jlab/clas/swimtools/Swim.java
@@ -1,7 +1,5 @@
 package org.jlab.clas.swimtools;
 
-import org.apache.commons.math3.util.FastMath;
-
 import org.jlab.geom.prim.Vector3D;
 import org.jlab.geom.prim.Point3D;
 import org.jlab.geom.prim.Line3D;
@@ -33,27 +31,8 @@ import org.jlab.clas.swimtools.Stoppers.ZSwimStopper;
  *
  * @author ziegler
  */
-public class Swim {
+public class Swim extends SwimPars implements ISwim {
 
-    final double SWIMZMINMOM = 0.75; // GeV/c
-    final double MINTRKMOM = 0.05; // GeV/c
-
-    private double _x0;
-    private double _y0;
-    private double _z0;
-    private double _phi;
-    private double _theta;
-    private double _pTot;
-    private final double _rMax = 5 + 3;
-    private double _maxPathLength = 9;
-    private boolean SwimUnPhys = false; //Flag to indicate if track is swimmable
-    private int _charge;
-    double accuracy = 20e-6; // 20 microns
-    public double stepSize = 5.00 * 1.e-4; // 500 microns
-    public double distanceBetweenSaves= 100*stepSize;
-    
-    private ProbeCollection PC;
-    
     /**
      * Class for swimming to various surfaces.  The input and output units are cm and GeV/c
      */
@@ -66,156 +45,12 @@ public class Swim {
     }
 
     /**
-     * Set max swimming path length
-     *
-     * @param _maxPathLength
-     */
-    public void setMaxPathLength(double _maxPathLength) {
-        this._maxPathLength = _maxPathLength;
-    }
-
-    /**
-     *
-     * @param direction +1 for out -1 for in
-     * @param x0 (cm)
-     * @param y0 (cm)
-     * @param z0 (cm)
-     * @param thx (units?)
-     * @param thy (units?)
-     * @param p (units?)
-     * @param charge
-     */
-    public void SetSwimParameters(int direction, double x0, double y0, double z0,
-                                  double thx, double thy, double p, int charge) {
-        _x0 = x0 / 100; // convert to meters
-        _y0 = y0 / 100;
-        _z0 = z0 / 100;
-        this.checkR(_x0, _y0, _z0);
-        double pz = direction * p / Math.sqrt(thx * thx + thy * thy + 1);
-        double px = thx * pz;
-        double py = thy * pz;
-        _phi = Math.toDegrees(FastMath.atan2(py, px));
-        _pTot = Math.sqrt(px * px + py * py + pz * pz);
-        _theta = Math.toDegrees(Math.acos(pz / _pTot));
-        _charge = direction * charge;
-    }
-
-    /**
-     * Sets the parameters used by swimmer based on the input track state vector
-     * parameters swimming outwards
-     *
-     * // z at a given DC plane in the tilted coordinate system
-     *
-     * @param superlayerIdx
-     * @param layerIdx
-     * @param x0 (cm)
-     * @param y0 (cm)
-     * @param z0 (cm)
-     * @param thx
-     * @param thy
-     * @param p
-     * @param charge
-     */
-    public void SetSwimParameters(int superlayerIdx, int layerIdx,
-                                  double x0, double y0, double z0,
-                                  double thx, double thy, double p, int charge) {
-        _x0 = x0 / 100; // convert to meters
-        _y0 = y0 / 100;
-        _z0 = z0 / 100;
-        this.checkR(_x0, _y0, _z0);
-        double pz = p / Math.sqrt(thx * thx + thy * thy + 1);
-        double px = thx * pz;
-        double py = thy * pz;
-        _phi = Math.toDegrees(FastMath.atan2(py, px));
-        _pTot = Math.sqrt(px * px + py * py + pz * pz);
-        _theta = Math.toDegrees(Math.acos(pz / _pTot));
-        _charge = charge;
-    }
-
-    /**
-     * Sets the parameters used by swimmer based on the input track parameters
-     *
-     * @param x0 (cm)
-     * @param y0 (cm)
-     * @param z0 (cm)
-     * @param px
-     * @param py
-     * @param pz
-     * @param charge
-     */
-    public void SetSwimParameters(double x0, double y0, double z0,
-                                  double px, double py, double pz, int charge) {
-        _x0 = x0 / 100; // convert to meters
-        _y0 = y0 / 100;
-        _z0 = z0 / 100;
-         this.checkR(_x0, _y0, _z0);
-        _phi = Math.toDegrees(FastMath.atan2(py, px));
-        _pTot = Math.sqrt(px * px + py * py + pz * pz);
-        _theta = Math.toDegrees(Math.acos(pz / _pTot));
-        _charge = charge;
-    }
-
-    /**
-     * 
-     * @param xcm
-     * @param ycm
-     * @param zcm
-     * @param phiDeg
-     * @param thetaDeg
-     * @param p
-     * @param charge
-     * @param maxPathLength
-     */
-    public void SetSwimParameters(double xcm, double ycm, double zcm,
-                                  double phiDeg, double thetaDeg,
-                                  double p, int charge, double maxPathLength) {
-        _maxPathLength = maxPathLength;
-        _charge = charge;
-        _phi = phiDeg;
-        _theta = thetaDeg;
-        _pTot = p;
-        _x0 = xcm / 100;
-        _y0 = ycm / 100;
-        _z0 = zcm / 100;
-        this.checkR(_x0, _y0, _z0);
-    }
-
-    /**
-     * 
-     * @param xcm
-     * @param ycm
-     * @param zcm
-     * @param phiDeg
-     * @param thetaDeg
-     * @param p
-     * @param charge
-     * @param maxPathLength
-     * @param Accuracy
-     * @param StepSize
-     */
-    public void SetSwimParameters(double xcm, double ycm, double zcm,
-                                  double phiDeg, double thetaDeg,
-                                  double p, int charge,
-                                  double maxPathLength, double Accuracy, double StepSize) {
-        _maxPathLength = maxPathLength;
-         accuracy = Accuracy/100;
-         stepSize = StepSize/100;
-        _charge = charge;
-        _phi = phiDeg;
-        _theta = thetaDeg;
-        _pTot = p;
-        _x0 = xcm / 100;
-        _y0 = ycm / 100;
-        _z0 = zcm / 100;
-        this.checkR(_x0, _y0, _z0);
-    }
-
-    /**
      * 
      * @param sector
      * @param z_cm
      * @return 
      */
+    @Override
     public double[] SwimToPlaneTiltSecSys(int sector, double z_cm) {
 
         // Fiducial Cut:
@@ -282,6 +117,7 @@ public class Swim {
      * @param z_cm
      * @return 
      */
+    @Override
     public double[] SwimToPlaneTiltSecSysBdlXZPlane(int sector, double z_cm) {
 
         // Fiducial Cut:
@@ -346,6 +182,7 @@ public class Swim {
      * @param z_cm
      * @return state  x,y,z,px,py,pz, pathlength, iBdl at the plane surface
      */
+    @Override
     public double[] SwimToPlaneLab(double z_cm) {
 
         // Fiducial Cut:
@@ -408,22 +245,10 @@ public class Swim {
 
     /**
      * 
-     * @param _x0
-     * @param _y0
-     * @param _z0 
-     */
-    private void checkR(double _x0, double _y0, double _z0) {
-        this.SwimUnPhys=false;
-        if(Math.sqrt(_x0*_x0 + _y0*_y0)>this._rMax || 
-                Math.sqrt(_x0*_x0 + _y0*_y0 + _z0*_z0)>this._maxPathLength)
-            this.SwimUnPhys=true;
-    }
-
-    /**
-     * 
      * @param Rad
      * @return state  x,y,z,px,py,pz, pathlength, iBdl at the surface 
      */
+    @Override
     public double[] SwimToCylinder(double Rad) {
         
         if (this.SwimUnPhys) return null;
@@ -463,6 +288,7 @@ public class Swim {
      * @param accuracy in cm 
      * @return state  x,y,z,px,py,pz, pathlength, iBdl at the surface 
      */
+    @Override
     public double[] SwimRho(double radius, double accuracy)  {
 
         if(this.SwimUnPhys) return null;
@@ -510,6 +336,7 @@ public class Swim {
      * @param accuracy in cm
      * @return swam trajectory to the cylinder
      */
+    @Override
     public double[] SwimGenCylinder(Point3D axisPoint1, Point3D axisPoint2, double radius, double accuracy)  {
 
         if(this.SwimUnPhys) return null;
@@ -556,6 +383,7 @@ public class Swim {
      * @param accuracy
      * @return 
      */
+    @Override
     public double[] SwimPlane(Vector3D n, Point3D p, double accuracy)  {
 
         if (this.SwimUnPhys) return null;
@@ -592,6 +420,7 @@ public class Swim {
      * @param Rad
      * @return state  x,y,z,px,py,pz, pathlength, iBdl at the surface 
      */
+    @Override
     public double[] SwimToSphere(double Rad) {
 
         if (this.SwimUnPhys==true) return null;
@@ -623,6 +452,7 @@ public class Swim {
      * @param dir
      * @return return state  x,y,z,px,py,pz, pathlength, iBdl at the plane surface in the lab frame
      */
+    @Override
     public double[] SwimToPlaneBoundary(double d_cm, Vector3D n, int dir) {
 
         if (this.SwimUnPhys) return null;
@@ -657,6 +487,7 @@ public class Swim {
         return value;
     }
 
+    @Override
     public double[] SwimToBeamLine(double xB, double yB) {
 
         if(this.SwimUnPhys==true) return null;
@@ -682,6 +513,7 @@ public class Swim {
         return value;
     }
    
+    @Override
     public double[] SwimToLine(Line3D l) {
         
         if (this.SwimUnPhys==true) return null;
@@ -707,43 +539,7 @@ public class Swim {
         return value;
     }
 
-    private void printV(String pfx, double v[]) {
-        double x = v[0] / 100;
-        double y = v[1] / 100;
-        double z = v[2] / 100;
-        double r = Math.sqrt(x * x + y * y + z * z);
-        System.out.println(String.format("%s: (%-8.5f, %-8.5f, %-8.5f) R: %-8.5f", pfx, z, y, z, r));
-    }
-
-    /**
-     * 
-     * @param sector
-     * @param x_cm
-     * @param y_cm
-     * @param z_cm
-     * @param result B field components in T in the tilted sector system
-     */
-    public void Bfield(int sector, double x_cm, double y_cm, double z_cm, float[] result) {
-        PC.RCP.field(sector, (float) x_cm, (float) y_cm, (float) z_cm, result);
-        result[0] = result[0] / 10;
-        result[1] = result[1] / 10;
-        result[2] = result[2] / 10;
-    }
-
-    /**
-     * 
-     * @param x_cm
-     * @param y_cm
-     * @param z_cm
-     * @param result B field components in T in the lab frame
-     */
-    public void BfieldLab(double x_cm, double y_cm, double z_cm, float[] result) {
-        PC.CP.field((float) x_cm, (float) y_cm, (float) z_cm, result);
-        result[0] = result[0] / 10;
-        result[1] = result[1] / 10;
-        result[2] = result[2] / 10;
-    }
-
+    @Override
     public double[] AdaptiveSwimPlane(double px, double py, double pz, double nx, double ny, double nz, double accuracy)  {
 
         if (this.SwimUnPhys) return null;
@@ -781,6 +577,7 @@ public class Swim {
         return value;
     }
     
+    @Override
     public double[] AdaptiveSwimCylinder(double a1x, double a1y, double a1z, double a2x, double a2y, double a2z, double radius, double accuracy)  {
         
         if (this.SwimUnPhys) return null;
@@ -821,6 +618,7 @@ public class Swim {
         return value;
     }
 
+    @Override
     public double[] AdaptiveSwimRho(double radius, double accuracy)  {
         System.out.println("Don't use yet");
         if(this.SwimUnPhys) return null;
@@ -859,6 +657,7 @@ public class Swim {
      * @param dir
      * @return state  x,y,z,px,py,pz, pathlength, iBdl at the surface 
      */
+    @Override
     public double[] SwimToZ(double Z, int dir) {
         double[] value = new double[8];
         ZSwimStopper stopper = new ZSwimStopper(Z, dir);
@@ -895,6 +694,7 @@ public class Swim {
         this.swimTraj = swimTraj;
     }
     
+    @Override
     public double[] SwimToDCA(SwimTrajectory trk2) { //use for both traj to get doca for each track
         
         double[] value = new double[6];
